res.data.cases instead of res.data in getCases (#50)
getCases response doesn't consist only of cases array, but starts with the following info:

{
  "offset": 0,
  "limit": 250,
  "size": 123,
  "_links": { "next": null, "prev": null },
  "cases": [

therefore in checking the length of res in index.js results in undefined and test run results are not published to TestRail. After this fix, the results are published again, as the getCases function returns only the cases array.